privesc-powershell
(WIP)

The above script is a PowerShell script that allows a user to choose from a list of system information commands. The script provides options to display information about the current operating system, the current hostname, the current username, or a list of user accounts on the system. Additionally, the script offers an option to run all the commands and output the results to a single text file. The output of each command is redirected to a file named "output.txt".